feat-#16: move models to domain
